fix: remove panics
